# internal/generator/rustrestgen/testdata/golden/fixture-api/src/router.rs
//! Code generated by psgen rustrestgen. DO NOT EDIT.
#![allow(unused_imports)]

use crate::interfaces::Implementations;
use axum::extract::{Path, Query, State};
use axum::http::{HeaderMap, Method};
use axum::routing::{delete, get, patch, post, put};
use axum::{Json, Router};
use psgen_http_runtime::{
    error_response, request_id_from_headers, wrap_envelope, RequestContext,
};
use serde_json::Value;
use std::collections::HashMap;
use std::sync::Arc;

#[derive(Clone)]
pub struct RouterState {
    pub implementations: Implementations,
}

pub fn build_router(implementations: Implementations) -> Router {
    let state = Arc::new(RouterState { implementations });
    let mut router: Router<Arc<RouterState>> = Router::new();
    router = router.route("/api/auth/me", get(handle_session_current_tenant));
    router = router.route("/api/tenant/custom-handler", post(handle_tenant_custom_handler));
    router = router.route("/api/tenants", get(handle_tenant_list_tenants));
    router = router.route("/api/tenants", post(handle_tenant_create_tenant));
    router = router.route("/api/tenants/{id}", get(handle_tenant_get_tenant));
    router = router.route("/api/tenants/{id}", patch(handle_tenant_update_secret));
    router.with_state(state)
}

fn method_from_str(method: &str) -> Method {
    match method {
        "get" => Method::GET,
        "post" => Method::POST,
        "put" => Method::PUT,
        "patch" => Method::PATCH,
        "delete" => Method::DELETE,
        _ => Method::POST,
    }
}

/// Convert an axum `HeaderMap` to a plain string map for `RequestContext`.
/// Header names are lowercased; non-UTF-8 values are dropped (per
/// EDR-0002 the SDK only forwards ASCII-safe headers anyway).
fn headers_to_map(headers: &HeaderMap) -> HashMap<String, String> {
    let mut out = HashMap::with_capacity(headers.len());
    for (name, value) in headers.iter() {
        if let Ok(v) = value.to_str() {
            out.insert(name.as_str().to_ascii_lowercase(), v.to_string());
        }
    }
    out
}
async fn handle_session_current_tenant(
    State(state): State<Arc<RouterState>>,
    headers: HeaderMap,
    Query(query): Query<HashMap<String, String>>,
) -> Result<Json<Value>, (axum::http::StatusCode, Json<Value>)> {
    let request_id = request_id_from_headers(&headers);
    let mut ctx = RequestContext::new(method_from_str("get"), "/api/auth/me".to_string());
    ctx.headers = headers_to_map(&headers);
    for (key, value) in query {
        ctx.query_params.insert(key, value);
    }
    let payload = Value::Null;
    let result = state
        .implementations
        .session
        .current_tenant(ctx, payload)
        .await;
    match result {
        Ok(body) => Ok(Json(wrap_envelope(body, request_id.as_deref()))),
        Err(err) => Err(error_response(err)),
    }
}
async fn handle_tenant_custom_handler(
    State(state): State<Arc<RouterState>>,
    headers: HeaderMap,
    Json(payload): Json<Value>,
) -> Result<Json<Value>, (axum::http::StatusCode, Json<Value>)> {
    let request_id = request_id_from_headers(&headers);
    let mut ctx = RequestContext::new(method_from_str("post"), "/api/tenant/custom-handler".to_string());
    ctx.headers = headers_to_map(&headers);
    let result = state
        .implementations
        .tenant
        .custom_handler(ctx, payload)
        .await;
    match result {
        Ok(body) => Ok(Json(wrap_envelope(body, request_id.as_deref()))),
        Err(err) => Err(error_response(err)),
    }
}
async fn handle_tenant_list_tenants(
    State(state): State<Arc<RouterState>>,
    headers: HeaderMap,
    Query(query): Query<HashMap<String, String>>,
) -> Result<Json<Value>, (axum::http::StatusCode, Json<Value>)> {
    let request_id = request_id_from_headers(&headers);
    let mut ctx = RequestContext::new(method_from_str("get"), "/api/tenants".to_string());
    ctx.headers = headers_to_map(&headers);
    for (key, value) in query {
        ctx.query_params.insert(key, value);
    }
    let payload = Value::Null;
    let result = state
        .implementations
        .tenant
        .list_tenants(ctx, payload)
        .await;
    match result {
        Ok(body) => Ok(Json(wrap_envelope(body, request_id.as_deref()))),
        Err(err) => Err(error_response(err)),
    }
}
async fn handle_tenant_create_tenant(
    State(state): State<Arc<RouterState>>,
    headers: HeaderMap,
    Json(payload): Json<Value>,
) -> Result<Json<Value>, (axum::http::StatusCode, Json<Value>)> {
    let request_id = request_id_from_headers(&headers);
    let mut ctx = RequestContext::new(method_from_str("post"), "/api/tenants".to_string());
    ctx.headers = headers_to_map(&headers);
    let result = state
        .implementations
        .tenant
        .create_tenant(ctx, payload)
        .await;
    match result {
        Ok(body) => Ok(Json(wrap_envelope(body, request_id.as_deref()))),
        Err(err) => Err(error_response(err)),
    }
}
async fn handle_tenant_get_tenant(
    State(state): State<Arc<RouterState>>,
    headers: HeaderMap,
    Path(path_params): Path<HashMap<String, String>>,
    Query(query): Query<HashMap<String, String>>,
) -> Result<Json<Value>, (axum::http::StatusCode, Json<Value>)> {
    let request_id = request_id_from_headers(&headers);
    let mut ctx = RequestContext::new(method_from_str("get"), "/api/tenants/{id}".to_string());
    ctx.headers = headers_to_map(&headers);
    for (key, value) in path_params {
        ctx.path_params.insert(key, value);
    }
    for (key, value) in query {
        ctx.query_params.insert(key, value);
    }
    let payload = Value::Null;
    let result = state
        .implementations
        .tenant
        .get_tenant(ctx, payload)
        .await;
    match result {
        Ok(body) => Ok(Json(wrap_envelope(body, request_id.as_deref()))),
        Err(err) => Err(error_response(err)),
    }
}
async fn handle_tenant_update_secret(
    State(state): State<Arc<RouterState>>,
    headers: HeaderMap,
    Path(path_params): Path<HashMap<String, String>>,
    Json(payload): Json<Value>,
) -> Result<Json<Value>, (axum::http::StatusCode, Json<Value>)> {
    let request_id = request_id_from_headers(&headers);
    let mut ctx = RequestContext::new(method_from_str("patch"), "/api/tenants/{id}".to_string());
    ctx.headers = headers_to_map(&headers);
    for (key, value) in path_params {
        ctx.path_params.insert(key, value);
    }
    let result = state
        .implementations
        .tenant
        .update_secret(ctx, payload)
        .await;
    match result {
        Ok(body) => Ok(Json(wrap_envelope(body, request_id.as_deref()))),
        Err(err) => Err(error_response(err)),
    }
}